feat(admin): group Slack Bot and Kilo Bot under Bots tabs
